M2-604: fix applet image
