+++ src/features/Library/Applet/AppletImage/AppletImage.styles.ts
@@ -3,14 +3,11 @@ import { StyledFlexAllCenter } from 'styles/styledComponents';
 
 import { variables } from 'styles/variables';
 
-export const StyledCustomCover = styled(StyledFlexAllCenter)`
+export const StyledAppletImageContainer = styled(StyledFlexAllCenter)`
   grid-row-start: 1;
   grid-row-end: 3;
   height: 10.4rem;
   width: 10.4rem;
-  border-radius: ${variables.borderRadius.md};
-  background-color: ${variables.palette.primary_container};
-  text-transform: uppercase;
 `;
 
 export const StyledImage = styled('img')`
@@ -18,4 +15,12 @@ export const StyledImage = styled('img')`
   height: 100%;
   object-fit: cover;
   border-radius: ${variables.borderRadius.md};
+`;
+
+export const StyledCustomCover = styled(StyledFlexAllCenter)`
+  width: 100%;
+  height: 100%;
+  border-radius: ${variables.borderRadius.md};
+  background-color: ${variables.palette.primary_container};
+  text-transform: uppercase;
 `;
